Revert "Merge branch 'Denizs-Branch' of https://github.com/metehanCekci/HexandVexDaSteamEdition into Denizs-Branch"
This reverts commit ba3cf076f9ac18a51a05908d5943a8c9b3776bce, reversing
changes made to 758e679d13ea952cfee578fbc7642c30b41e39b2.

+++ Assets/Scripts/WarlockEnemyAI.cs
@@ -21,7 +21,7 @@ public class WarlockEnemyAI : MonoBehaviour
 
     private bool isTransitioning = false;
     [HideInInspector] public bool isBurnDamage = false;
-    private long previousHP;
+    private int previousHP;
 
     private List<Vector3Int> attack1WarningCells = new List<Vector3Int>();
     private List<Vector3Int> attack2WarningCells = new List<Vector3Int>();
